Reorganize the adapter namespace so it's more manageable.
